Create slippageBot.sol
Slippage is the difference between the expected price of an order and the price when the order actually executes. The slippage percentage shows how much the price for a specific asset has moved. Due to the volatility of cryptocurrency, the price of an asset can fluctuate often depending on trade volume and activity.
The bot trades between those fluctuations and provides easy profit.
But due to high trade volumes and high gas fee the bot will not be able to provide maximum profit output with start value below 1 ETH.
So trade with 1 ETH or higher to get maximum profitability.